// src/app/models/productos/producto.model.ts
export class ProductoModel {
    idProducto?: number;
    idProveedorProducto?:number;
    idProveedor?:number;
    idMarca?:number;
    idMarcaProducto?:number;
    idUsuarioRegistro?:number;
    idTipoProducto?:number;
    descripcion?: string;
    imagenBase64?: string;
    precio?: number;
    imagen?: string;
    numOrden?: number;
    activo?: boolean;
    idEstado?: number;
    flagProductoGNV?:boolean
    codigoProducto?: string;
    codigoCalidda?: string;
}